editor/libeditor/tests/test_resizers_appearance.html

<!DOCTYPE html>
<html>
<head>
  <title>Test for resizers appearance</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none;">

</div>

<div id="editor" contenteditable></div>
<div id="clickaway" style="width: 3px; height: 3px;"></div>
<img src="green.png"><!-- for ensuring to load the image at first test of <img> case -->
<pre id="test">

<script class="testbody" type="application/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(async function() {
  async function waitForSelectionChange() {
    return new Promise(resolve => {
      document.addEventListener("selectionchange", () => {
        resolve();
      }, {once: true});
    });
  }

  let editor = document.getElementById("editor");
  let outOfEditor = document.getElementById("clickaway");

  const kTests = [
    { description: "<img>",
      innerHTML: "<img id=\"target\" src=\"green.png\" width=\"100\" height=\"100\">",
      resizable: true,
    },
    { description: "<table>",
      innerHTML: "<table id=\"target\" border><tr><td>1-1</td><td>1-2</td></tr><tr><td>2-1</td><td>2-2</td></tr></table>",
      resizable: true,
    },
    { description: "absolute positioned <div>",
      innerHTML: "<div id=\"target\" style=\"position: absolute; top: 50px; left: 50px;\">positioned</div>",
      resizable: true,
    },
    { description: "fixed positioned <div>",
      innerHTML: "<div id=\"target\" style=\"position: fixed; top: 50px; left: 50px;\">positioned</div>",
      resizable: false,
    },
    { description: "relative positioned <div>",
      innerHTML: "<div id=\"target\" style=\"position: relative; top: 50px; left: 50px;\">positioned</div>",
      resizable: false,
    },
  ];

  for (const kTest of kTests) {
    const kDescription = kTest.description + ": ";
    editor.innerHTML = kTest.innerHTML;
    let target = document.getElementById("target");

    document.execCommand("enableObjectResizing", false, false);
    ok(!document.queryCommandState("enableObjectResizing"),
       kDescription + "Object resizer should be disabled by the call of execCommand");

    synthesizeMouseAtCenter(outOfEditor, {});
    let promiseSelectionChangeEvent1 = waitForSelectionChange();
    synthesizeMouseAtCenter(target, {});
    await promiseSelectionChangeEvent1;

    ok(!target.hasAttribute("_moz_resizing"),
       kDescription + ": While enableObjectResizing is disabled, resizers shouldn't appear");

    document.execCommand("enableObjectResizing", false, true);
    ok(document.queryCommandState("enableObjectResizing"),
       kDescription + "Object resizer should be enabled by the call of execCommand");

    synthesizeMouseAtCenter(outOfEditor, {});
    let promiseSelectionChangeEvent2 = waitForSelectionChange();
    synthesizeMouseAtCenter(target, {});
    await promiseSelectionChangeEvent2;

    is(target.hasAttribute("_moz_resizing"), kTest.resizable,
       kDescription + (kTest.resizable ? "While enableObjectResizing is enabled, resizers should appear" :
                                         "Even while enableObjectResizing is enabled, resizers shouldn't appear"));
  }

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
